Fix for fatal on transform of empty section heading
Jon and I were able to track the cause of this down to empty section
headings, which sometimes get generated via templates, for instance
above the references in this page:

https://pt.m.wikipedia.org/w/index.php?oldid=46713112&title=Theatro_S%C3%A3o_Pedro_(S%C3%A3o_Paulo)

This suggested the proper fix is to go ahead and leave `header.firstChild`
in place as the relative insertion target, as if it's `null` this will be
interpreted as appending at the end (of an empty set). This properly puts
the toggle click widget in the proper place if one of these headings
shows up with no heading text, without failing as previously.

Bug: T349317
Change-Id: Ifb3d1f1a62098eb3b444b126e5ad35fdc5fbe89e