tests: change ApiParseExtenderTest to 'fallback' skin
When MinervaNeue is not available, the test will fail because
`useskin=minerva` is not a valid value.  ApiParse needs a skin to be
passed in order to switch to skin mode, the `fallback` skin is provided
by MediaWiki core and is always available.

Change-Id: I2104672dc250071c8d6287c43d8a5f4634fa998a
(cherry picked from commit 8f7a8bf15bff2b9a1599564e135543fd97b7cca9)