UADeviceDetector: Avoid "samsung" device token matching "SamsungBrowser"
This token is here to match legacy device model tokens. Those devices
ran various mobile browsers that we do not know or match the names of
(e.g. "Jasmine" [1]), and by matching the device model token we easily
classify them all as mobile.

Since the introduction of "Samsung Internet", a Chrome-based browser
for Samsung devices running Android, we don't need to match the device
token as we already detect these by their operating system (Android)
plus the browser engine also tells (Chrome "Mobile", "Mobile" Firefox).

While it seems harmless to match, it actually poses a problem because
SamsungBrowser has a desktop mode as well. This is used by system
configuration, such as the case for Samsung SMART-TV devices, and can
also be used via user choice via the "Prefer desktop websites" setting
or the "Desktop site" button on a per-site basis. In desktop mode,
Samsung Internet swaps the "Android" OS token for a desktop "Linux"
token, and keeps the "SamsungBrowser" app name, which seems fair.
But, our regex was misclassifying those as mobile browsers.

There was a workaround for the SMART-TV case already, which is
redundant after this patch because it wasn't actually a special case
in terms of its user agent string. SamsungBrowser on SMART-TV is
acting the same way as Chrome or Firefox in desktop mode, with a
desktop "Chrome/X.Y" token (not "Chrome/X.Y Mobile"), and a
desktop "Linux" token (not "Android").

Both the legacy SAMSUNG/Jasmine case (since 2013 with I79401aee6ec9 for
T32827), and the SMART-TV case have existing test cases that continue
to pass.

[1]: https://en.wikipedia.org/wiki/Samsung_GT-S5230
[2]: https://www.chromium.org/updates/ua-reduction/

Bug: T405279
Change-Id: I3e4f0fb5c01f6fb84a93da8128adcadbd7d3b2c6