Removal of AMC outreach drawer, MobileFrontend side
The toggle switch on the options continues to work, but the separate
callout should now be silent.

Appears to work on quick check, and is passing tests.
Can cause errors in some event handlres if the Minerva side is not
applied at the same time or before this, so before to grab the
dependent patch there when backporting or deploying.

Bug: T402470
Change-Id: I3cadc7784165328d9021fda685574c05a6ee07f2
Depends-On: I3a543e26bcb3ad155912c94700c151ba34aea66d